Identity service

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -53,13 +54,14 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
             var auction = _mapper.Map<Auction>(auctionDto);
 
             // TODO: add current user as seller
-            auction.Seller = "Test";
+            auction.Seller = User.Identity.Name;
 
             _context.Auctions.Add(auction);
 
@@ -77,6 +79,7 @@ namespace AuctionService.Controllers
             return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, newAuction);
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
@@ -86,6 +89,11 @@ namespace AuctionService.Controllers
             if (auction == null)
             {
                 return NotFound();
+            }
+
+            if (auction.Seller != User.Identity.Name)
+            {
+                return Forbid();
             }
 
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make; 
@@ -106,6 +114,7 @@ namespace AuctionService.Controllers
             return BadRequest("Problem saving changes");
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
@@ -117,7 +126,10 @@ namespace AuctionService.Controllers
                 return NotFound();
             }
 
-            // TODO: Check seller == username
+            if (auction.Seller != User.Identity.Name)
+            {
+                return Forbid();
+            }
 
             _context.Auctions.Remove(auction);
